@@ -6,7 +6,10 @@
     "autoload": {
         "psr-4": {
             "Externals\\": "src/"
-        }
+        },
+        "files": [
+            "init.php"
+        ]
     },
     "autoload-dev": {
         "psr-4": {
@@ -14,33 +17,39 @@
         }
     },
     "require": {
-        "php": "^7.4",
-        "php-di/php-di": "^6.0",
-        "stratify/framework": "~0.7.1",
-        "twig/twig": "^2.0",
-        "mnapoli/silly": "^1.7",
-        "doctrine/dbal": "^2.10",
-        "twig/extensions": "^1.5",
-        "vstelmakh/url-highlight": "^3.0",
-        "monolog/monolog": "^1.25",
-        "symfony/monolog-bridge": "^3.4",
-        "psr7-sessions/storageless": "^2.0",
+        "php": "^8.0",
+        "ext-dom": "*",
+        "ext-json": "*",
+        "ext-pdo": "*",
+        "algolia/algoliasearch-client-php": "^1.28",
+        "bref/bref": "^1.2",
+        "bref/logger": "^1.0",
+        "bref/micro": "^0.1.0",
+        "doctrine/dbal": "^3.0",
         "league/commonmark": "^1.3",
-        "gravatarphp/gravatar": "^2.0",
-        "gravatarphp/twig-integration": "^2.0",
         "league/oauth2-client": "^2.5",
         "league/oauth2-github": "^2.0",
-        "wizacha/new-relic-polyfill": "^1.0",
-        "algolia/algoliasearch-client-php": "^1.28",
+        "mnapoli/silly": "^1.7",
+        "php-di/php-di": "^7.0.0-beta2",
+        "psr/http-message": "^1.0",
+        "psr/http-server-handler": "^1.0",
+        "psr/http-server-middleware": "^1.0",
+        "psr7-sessions/storageless": "^7.0",
         "rvdv/nntp": "^0.9.2",
-        "zbateson/mail-mime-parser": "^0.4.13",
-        "vlucas/phpdotenv": "^2.6",
-        "sentry/sentry": "^1.11",
-        "bref/bref": "1.0.0-beta2",
-        "bref/logger": "^1.0"
+        "sentry/sdk": "^3.0",
+        "twig/extensions": "^1.5",
+        "twig/twig": "^2.0",
+        "vlucas/phpdotenv": "^5.0",
+        "vstelmakh/url-highlight": "^3.0",
+        "zbateson/mail-mime-parser": "^1.0"
     },
     "require-dev": {
+        "bref/dev-server": "^0.2",
+        "mnapoli/hard-mode": "^0.3",
         "phpunit/phpunit": "^9.2.6",
-        "mnapoli/hard-mode": "^0.1.1"
+        "vimeo/psalm": "^4.2"
+    },
+    "config": {
+        "sort-packages": true
     }
 }
