@@ -13,7 +13,7 @@ use Doctrine\DBAL\Platforms\MySQL57Platform;
  */
 class CustomMySQLPlatform extends MySQL57Platform
 {
-    protected function _getCreateTableSQL($tableName, array $columns, array $options = []): array
+    protected function _getCreateTableSQL($name, array $columns, array $options = []): array
     {
         // Charset
         if (! isset($options['charset'])) {
@@ -25,6 +25,6 @@ class CustomMySQLPlatform extends MySQL57Platform
             $options['collate'] = 'utf8mb4_unicode_ci';
         }
 
-        return parent::_getCreateTableSQL($tableName, $columns, $options);
+        return parent::_getCreateTableSQL($name, $columns, $options);
     }
 }
