@@ -5,20 +5,23 @@ namespace Externals\Application\Middleware;
 use Externals\User\User;
 use Psr\Http\Message\ResponseInterface;
 use Psr\Http\Message\ServerRequestInterface;
-use PSR7Session\Http\SessionMiddleware;
-use Stratify\Http\Middleware\Middleware;
+use Psr\Http\Server\MiddlewareInterface;
+use Psr\Http\Server\RequestHandlerInterface;
+use PSR7Sessions\Storageless\Http\SessionMiddleware;
+use PSR7Sessions\Storageless\Session\SessionInterface;
 
 /**
  * Set the user in the request.
  */
-class AuthMiddleware implements Middleware
+class AuthMiddleware implements MiddlewareInterface
 {
-    public function __invoke(ServerRequestInterface $request, callable $next): ResponseInterface
+    public function process(ServerRequestInterface $request, RequestHandlerInterface $handler): ResponseInterface
     {
         $session = $request->getAttribute(SessionMiddleware::SESSION_ATTRIBUTE);
+        assert($session instanceof SessionInterface);
 
         $request = $request->withAttribute('user', User::fromData($session->get('user')));
 
-        return $next($request);
+        return $handler->handle($request);
     }
 }
